Hack to partially fix mono depth buffer in FC4
On every draw call, this checks to see if the known bad depth texture is
in use. If it is, it removes the depth buffer as an active render target
so that it may be used as an input to the pixel shader and replaces the
bad depth texture with the ZRepair one (which hopefully will be the
active depth buffer).

This works since all the broken effects will render transparent if a
depth test in the pixel shader fails, so skipping the depth test before
running the pixel shader is ok. There might still be some cases where
this is not the case - I get dark rectangles appearing indoors that I
haven't seen before, but I also haven't fixed all the remaining issues
in those areas so that may be unrelated.

Currently only works on MSAA/TXAA - in part because the depth buffer
hashes are hardcoded as the active ones on those AA settings, and
partially because this doesn't seem to work properly on no AA or SMAA as
the depth buffer seems to still be mono in that case (but maybe I've
just grabbed the wrong hash).

Unfortunately this technique seems to break the automatic stereo
adjustment applied by the driver. I've manually applied it to half a
dozen vertex shaders as proof of concept, but the problem is it is
broken inconsistently leading to massive flickering depending on the
camera angle. Possibly related - it seems that the ZRepair texture may
not always be valid as the water can appear fully opaque depending on
camera angle.

It's so inconsistent it can lead to a seam in the water between where it
is using the correct stereo correction and depth buffer and where it is
not.

Still - progress.